+++ Gruntfile.js
@@ -8,39 +8,29 @@ module.exports = function(grunt) {
 		},
 		jshint: {
 			devel: {
-				options: {
-					asi: false,
-					browser: true,
-					devel: true,
-					debug: true
-				},
-				src: ['src/*.js', '!src/template.js']
+				options: grunt.file.readJSON("./components/project-settings/jshint-dev.json"),
+				src: ['src/**/*.js']
 			},
 			release: {
-				options: {
-					browser: true
-				},
-				src: ['src/*.js', '!src/template.js']
-			}
-		},
-		uglify: {
-			all: {
-				src: "dist/<%=pkg.name%>.js",
-				dest: "dist/<%=pkg.name%>.min.js"
-			},
-			amd: {
-				src: ['dist/amd.mtvn.js'],
-				dest: 'dist/amd.mtvn.min.js'
+				options: grunt.file.readJSON("./components/project-settings/jshint.json"),
+				src: ['src/**/*.js']
 			}
 		},
 		rig: {
 			devel: {
-				src: ['src/build/<%= pkg.name %>.js'],
-				dest: 'dist/<%= pkg.name %>.js'
-			},
-			amd: {
-				src: ['src/build/amd.mtvn.js'],
-				dest: 'dist/amd.mtvn.js'
+				expand: true,
+				cwd: "src/build/",
+				src: '*.js',
+				dest: 'dist/'
+			}
+		},
+		uglify: {
+			devel: {
+				expand: true,
+				cwd: "dist/",
+				src: '*.js',
+				ext: ".min.js",
+				dest: 'dist/'
 			}
 		},
 		handlebars: {
@@ -80,6 +70,9 @@ module.exports = function(grunt) {
 					}, {
 						match: 'version',
 						replacement: '<%= pkg.version %><%= grunt.config("buildNumber") %>'
+					}, {
+						match: 'package-name',
+						replacement: '<%= pkg.name %>'
 					}]
 				},
 				src: "dist/*.js",