+++ src/controls/controls.js
@@ -1,19 +1,29 @@
 /* exported Controls */
-/* global _, Backbone, $, Events, Slider, PlayPauseButton, VolumeButton, ClosedCaptionButton, Templates*/
+/* global _, Backbone, $, Events, Slider, PlayPauseButton, 
+  LiveButton, VolumeButton, ClosedCaptionButton, Templates*/
 var Controls = (function() {
+	//= slider/slider.js
 	var CONTROLS_TEMPLATE = Templates["src/controls/template.html"],
+
 		css = {
 			hide: "mtvn-controls-hidden",
 			slider: "mtvn-controls-slider",
 			playPause: "mtvn-controls-play-pause",
+			live: "mtvn-controls-live",
 			volume: "mtvn-controls-volume",
 			cc: "mtvn-controls-cc"
+		},
+		addEvents = function(listener, dispatcher, events, cb) {
+			_.each(events, function(eventName) {
+				listener.listenTo(dispatcher, eventName, cb);
+			});
 		};
 	return Backbone.View.extend({
 		tagName: "div",
 		className: "mtvn-controls",
 		events: {
-			"click .mtvn-controls-fullscreen": "onFullscreen"
+			"click .mtvn-controls-fullscreen": "onFullscreen",
+			"click .mtvn-controls-rewind": "onRewind"
 		},
 		initialize: function(options) {
 			this.options = options;
@@ -31,31 +41,43 @@ var Controls = (function() {
 				el: this.$("." + css.playPause),
 				paused: options.paused
 			});
-			this.listenTo(this.playPauseButton, Events.PLAY, this.sendEvent);
-			this.listenTo(this.playPauseButton, Events.PAUSE, this.sendEvent);
+			addEvents(this, this.playPauseButton, [Events.PLAY, Events.PAUSE], this.sendEvent);
 			// SLIDER
 			this.slider = new Slider({
 				el: this.$("." + css.slider),
 				playhead: options.playhead,
+				isLive: options.live,
 				durations: options.durations
 			});
+			// Seek Event
 			this.listenTo(this.slider, Events.SEEK, this.sendEvent);
+			// LIVE
+			if (options.live || options.isLive) {
+				this.liveButton = new LiveButton({
+					el: this.$("." + css.live),
+					isLive: options.live
+				});
+				// Live Event
+				this.listenTo(this.liveButton, Events.GO_LIVE, this.sendEvent);
+				// Handle IS_LIVE Event
+				this.liveButton.listenTo(this.slider, Events.IS_LIVE, this.liveButton.onLiveChange);
+			}
 			// VOLUME
 			if (options.showVolume) {
 				this.volumeButton = new VolumeButton({
 					volume: options.volume,
+					showVolumeSlider: options.showVolumeSlider,
 					el: this.$("." + css.volume)
 				});
-				this.listenTo(this.volumeButton, Events.VOLUME, this.sendEvent);
-				this.listenTo(this.volumeButton, Events.MUTE, this.sendEvent);
-				this.listenTo(this.volumeButton, Events.UNMUTE, this.sendEvent);
+				// Volume Events
+				addEvents(this, this.volumeButton, [Events.VOLUME, Events.MUTE, Events.UNMUTE], this.sendEvent);
 			}
-
 			// CC
 			this.closedCaptionButton = new ClosedCaptionButton({
 				ccEnabled: options.ccEnabled,
 				el: this.$("." + css.cc)
 			});
+			// CC Event
 			this.listenTo(this.closedCaptionButton, Events.CC, this.sendEvent);
 		},
 		hide: function() {
@@ -80,6 +102,7 @@ var Controls = (function() {
 			this.playPauseButton.setPaused(paused);
 		},
 		getPlayhead: function() {
+			// used for testing.
 			return this.slider.playhead;
 		},
 		setPlayhead: function(playhead) {
@@ -94,6 +117,11 @@ var Controls = (function() {
 		sendEvent: function(event) {
 			event.target = this;
 			this.trigger(event.type, event);
+		},
+		onRewind: function() {
+			this.sendEvent({
+				type: Events.REWIND
+			});
 		},
 		onFullscreen: function() {
 			this.sendEvent({